+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.4
+
+* update style custom for text item
+* update style custom for text label
+
+## 0.0.3
+
+* update public API has dartdoc comments
+
+## 0.0.2
+
+* update screenshots
+
 ## 0.0.1
 
 * Init release